refactor(models): Rename folder with dedicated model files
