+++ app/database/sessions.py
@@ -1,5 +1,5 @@
 from flask import g, Flask
-from app.model.database import SessionLocal
+from .base import SessionLocal
 from sqlalchemy.orm import Session
 from typing import Optional
