+++ app/models/__init__.py
@@ -1,0 +1,4 @@
+from .users import User
+from .products import Product
+from .orders import Order
+from .items import OrderItem
