+++ app/routes/main_routes.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, jsonify, g, Response
 from app.core.permissions import auth_required
-from app.model.models import User
+from app.models import User
 from typing import Tuple
 
 main_bp = Blueprint("main", __name__)
