+++ app/services/product_services.py
@@ -1,4 +1,4 @@
-from app.model.models import Product
+from app.models import Product
 from app.core.exceptions.app_errors import NotFoundError
 from sqlalchemy.orm import Session
 from typing import List, Optional
