+++ app/tests/conftest.py
@@ -1,8 +1,8 @@
 import pytest
 from flask import g
 from sqlalchemy.orm import Session
-from app.model.database import SessionLocal
-from app.model.models import User, Product, Order, OrderItem
+from app.database.base import SessionLocal
+from app.models import User, Product, Order, OrderItem
 from app.core.auth_utils import generate_token
 from werkzeug.security import generate_password_hash
 from app.services.product_services import add_product
@@ -12,7 +12,7 @@ from flask.testing import FlaskClient
 from typing import Tuple, Generator, List, Dict, Any
 from app import create_app
 import os
-from app.core.db_manager import DatabaseManager
+from app.database.db_manager import DatabaseManager
 
 # Force TestConfig à la creation de app
 os.environ["FLASK_ENV"] = "testing"
