+++ app/tests/test_users.py
@@ -2,14 +2,14 @@ import pytest
 import jwt
 from datetime import datetime
 from werkzeug.security import check_password_hash
-from app.model.models import User
-from app.config import Config
+from app.models import User
+# from app.config import Config
 from app.core.auth_utils import get_user_by_email
 
 from typing import Tuple, Dict
 from flask.testing import FlaskClient
 from sqlalchemy.orm import Session
-from flask import current_app
+# from flask import current_app
 
 from app.core.auth_utils import jwt_settings
